@@ -1,23 +1,25 @@
 # frozen_string_literal: true
 # Users::SessionsController
-class Users::SessionsController < Devise::SessionsController
-  # before_filter :configure_sign_in_params, only: [:create]
-  skip_before_action :authenticate_user!
+module Users
+  class SessionsController < Devise::SessionsController
+    # before_filter :configure_sign_in_params, only: [:create]
+    skip_before_action :authenticate_user!
 
-  # GET /resource/sign_in
-  # def new
-  #   super
-  # end
+    # GET /resource/sign_in
+    # def new
+    #   super
+    # end
 
-  # DELETE /resource/sign_out
-  # def destroy
-  #   super
-  # end
+    # DELETE /resource/sign_out
+    # def destroy
+    #   super
+    # end
 
-  protected
+    # protected
 
-  # If you have extra params to permit, append them to the sanitizer.
-  # def configure_sign_in_params
-  #   devise_parameter_sanitizer.for(:sign_in) << :attribute
-  # end
+    # If you have extra params to permit, append them to the sanitizer.
+    # def configure_sign_in_params
+    #   devise_parameter_sanitizer.for(:sign_in) << :attribute
+    # end
+  end
 end
