@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 # EventsUser
-class EventsUser < ActiveRecord::Base
+class EventsUser < ApplicationRecord
   belongs_to :user
   belongs_to :event
 end
