@@ -1,20 +1,16 @@
-# frozen_string_literal: true
-require File.expand_path("../boot", __FILE__)
+require_relative "boot"
 
 require "sprockets/es6"
 require "rails/all"
 
 # Require the gems listed in Gemfile, including any gems
-# you"ve limited to :test, :development, or :production.
+# you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
 module Cal4w
   class Application < Rails::Application
-    # Use the responders controller from the responders gem
-    config.app_generators.scaffold_controller :responders_controller
-
-    # Settings in config/environments/* take precedence over those specified
-    # here.
+    # Settings in config/environments/* take precedence over
+    # those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
@@ -24,16 +20,9 @@ module Cal4w
     # Default is UTC.
     config.time_zone = "Seoul"
 
-    # The default locale is :en and all translations from
-    # config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += \
-    # Dir[Rails.root.join("my locales", "*.{rb,yml}").to_s]
-    # config.i18n.default_locale = :de
-
     config.autoload_paths << Rails.root.join("lib")
     config.autoload_paths << Rails.root.join("app/forms")
 
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    config.eager_load_paths << Rails.root.join("lib")
   end
 end
