@@ -42,77 +42,77 @@ RSpec.describe EventsController, type: :controller do
     sign_in @user
   end
 
-  describe 'GET #index' do
+  describe "GET #index" do
     it "return http status 200" do
-      get :index, {}
+      get :index
       expect(response).to have_http_status(:success)
     end
   end
 
-  describe 'GET #show' do
+  describe "GET #show" do
     it "assigns the requested event as @event" do
       event = create(:event)
-      get :show, id: event.to_param
+      get :show, params: { id: event.to_param }
       expect(assigns(:event)).to eq(event)
     end
   end
 
-  describe 'POST #join' do
+  describe "POST #join" do
     it "creates a new Member" do
       event = create(:event)
       expect do
-        post :join, id: event.to_param
+        post :join, params: { id: event.to_param }
       end.to change(event.members, :count).by(1)
     end
   end
 
-  describe 'POST #unjoin' do
+  describe "POST #unjoin" do
     it "creates a new Member" do
       event = create(:event)
       event.members << @user
       expect do
-        post :unjoin, id: event.to_param
+        post :unjoin, params: { id: event.to_param }
       end.to change(event.members, :count).by(-1)
     end
   end
 
-  describe 'GET #edit' do
+  describe "GET #edit" do
     it "will be redirected root_path if try to edit the others event" do
       event = create(:event)
-      get :edit, id: event.to_param
+      get :edit, params: { id: event.to_param }
       expect(response).to redirect_to("/")
     end
   end
 
-  describe 'POST #create' do
+  describe "POST #create" do
     context "with valid params" do
       it "creates a new Event" do
         expect do
-          post :create, event_form: valid_attributes
+          post :create, params: { event_form: valid_attributes }
         end.to change(Event, :count).by(1)
       end
 
       it "assigns a newly created event as @event" do
-        post :create, event_form: valid_attributes
+        post :create, params: { event_form: valid_attributes }
         expect(assigns(:event)).to be_a(Event)
         expect(assigns(:event)).to be_persisted
       end
 
       it "redirects to the created event" do
-        post :create, event_form: valid_attributes
+        post :create, params: { event_form: valid_attributes }
         expect(response).to redirect_to(Event.last)
       end
     end
 
     context "with invalid params" do
       it "re-renders the 'new' template" do
-        post :create, event_form: invalid_attributes
+        post :create, params: { event_form: invalid_attributes }
         expect(response).to render_template("new")
       end
     end
   end
 
-  describe 'PUT #update' do
+  describe "PUT #update" do
     context "with valid params" do
       let(:new_attributes) do
         { subject: "New Subject" }
@@ -120,20 +120,22 @@ RSpec.describe EventsController, type: :controller do
 
       it "updates the requested event" do
         event = create(:event, user: @user)
-        put :update, id: event.to_param, event_form: new_attributes
+        put :update, params: { id: event.to_param, event_form: new_attributes }
         event.reload
         expect(assigns(:event)).to eq(event)
       end
 
       it "assigns the requested event as @event" do
         event = create(:event, user: @user)
-        put :update, id: event.to_param, event_form: valid_attributes
+        put :update,
+            params: { id: event.to_param, event_form: valid_attributes }
         expect(assigns(:event)).to eq(event)
       end
 
       it "redirects to the event" do
         event = create(:event, user: @user)
-        put :update, id: event.to_param, event_form: valid_attributes
+        put :update,
+            params: { id: event.to_param, event_form: valid_attributes }
         expect(response).to redirect_to(event)
       end
     end
@@ -141,23 +143,24 @@ RSpec.describe EventsController, type: :controller do
     context "with invalid params" do
       it "re-renders the 'edit' template" do
         event = create(:event, user: @user)
-        put :update, id: event.to_param, event_form: invalid_attributes
+        put :update,
+            params: { id: event.to_param, event_form: invalid_attributes }
         expect(response).to render_template("edit")
       end
     end
   end
 
-  describe 'DELETE #destroy' do
+  describe "DELETE #destroy" do
     it "destroys the requested event" do
       event = create(:event, user: @user)
       expect do
-        delete :destroy, id: event.to_param
+        delete :destroy, params: { id: event.to_param }
       end.to change(Event, :count).by(-1)
     end
 
     it "redirects to the events list" do
       event = create(:event, user: @user)
-      delete :destroy, id: event.to_param
+      delete :destroy, params: { id: event.to_param }
       expect(response).to redirect_to(events_url)
     end
   end
